+++ RPG Demo/RPG/SceneMain.cpp
@@ -135,8 +135,8 @@ bool cSceneMain::collide() {
 void cSceneMain::render() {
     //if (KEY_DOWN(VK_ESCAPE))
     //    selectedRole->render();
-    for (int i = 0; i < m_pCurentMap->m_iRow; i++) {
-        for (int j = 0; j < m_pCurentMap->m_iCol; j++) {
+    for (size_t i = 0; i < m_pCurentMap->m_iRow; i++) {
+        for (size_t j = 0; j < m_pCurentMap->m_iCol; j++) {
             int iTag = 0;
             if (m_pNPCRuler->render(i, j, m_iMapID)/*m_pCurentMap->m_arrMap[i][j] >= 400 && m_pCurentMap->m_arrMap[i][j] < 500*/) {
                 //render函数 判断-返回 要打印的字符
